Add files via upload
Lots of little bug fixes. Particularly getting the sequence to capture correctly. Also fixed were some issues around the Exposure mode /  analogue digital gain. If the -ex off tag is present  it appear that the -ag and -dg values are not used. Now when -ag and -dg are being used no -ex or -iso tags are used. Some more bugs and typos fixed along with some real world testing. Seems to be working in the sequence capture mode.